Refactor members table and spinner.
