+++ client/app/bundles/Events/components/MembersTable.jsx
@@ -1,0 +1,108 @@
+import React, { Component } from 'react';
+
+import Member from './Member';
+import SortableHeader from './SortableHeader';
+import Spinner from './Spinner';
+
+class MembersTable extends Component {
+  state = { emails: [] };
+
+  groupColumn() {
+    if (this.props.showGroupName)
+      return <SortableHeader title='Group Name' sortBy='group_name'  style={{ width: '20%' }}/>
+  }
+
+  locationColumn() {
+    const width = this.props.showGroupName ? '10%' : '20%'
+    return <th style={{ width: `${width}`}}>Location</th>
+  }
+
+  phoneColumn() {
+    const width = this.props.showGroupName ? '10%' : '15%'
+    return <th style={{ width: `${width}`}}>Phone</th>
+  }
+
+  nameColumn() {
+    const width = this.props.showGroupName ? '15%' : '20%'
+    return <SortableHeader title='Name' sortBy='name' style={{ width: `${width}`}} />
+  }
+
+  generateEmailsLink() {
+    const { emails } = this.state;
+    if (emails.length)
+      return (
+        <a href={`mailto:${emails.join(',')}`} className='fa fa-envelope-o'/>
+      );
+    else
+      return (<i className='fa fa-envelope-o'/>);
+  }
+
+  addMemberEmail(email) {
+    const emails = this.state.emails.concat(email);
+    this.setState({ emails });
+  }
+
+  removeMemberEmail(memberEmail) {
+    const emails = _.filter(this.state.emails, (email) => (email !== memberEmail));
+    this.setState({ emails });
+  }
+
+  renderMembers() {
+    return this.props.memberships.map(membership => {
+      const { tags } = membership.attributes;
+      const person = membership.attributes.person.data;
+      return <Member key={person.id} id={person.id}
+        member={person.attributes}
+        groups={person.relationships.groups.data}
+        role={membership.attributes.role}
+        onCheckboxChecked={this.addMemberEmail.bind(this)}
+        onCheckboxUnChecked={this.removeMemberEmail.bind(this)}
+        showGroupName={this.props.showGroupName}
+        membershipId={membership.id}
+        tags={tags}
+      />
+    })
+  }
+
+  render() {
+    const { memberships } = this.props;
+
+    if (!memberships)
+      return (
+        <div style={{ alignItems: 'center', height: '150px', display: 'flex', justifyContent: 'center' }}>
+          <Spinner classes='fa-3x' />
+        </div>
+      )
+
+    if (!memberships.length)
+      return (
+        <div style={{ justifyContent: 'center', height: '150px', display: 'flex', alignItems: 'center' }}>
+          <h4>
+            No members found
+          </h4>
+        </div>
+      );
+
+    return (
+      <table className='table'>
+        <thead>
+          <tr>
+            <th style={{ width: '5%'}}>{this.generateEmailsLink()}</th>
+            {this.nameColumn()}
+            {this.phoneColumn()}
+            {this.locationColumn()}
+            {this.groupColumn()}
+            <th style={{ width: '20%'}}>Tags</th>
+            <SortableHeader title='Role' sortBy='role' style={{ width: '15%'}} />
+            <th style={{ width: '5%'}}></th>
+          </tr>
+        </thead>
+        <tbody>
+          {this.renderMembers()}
+        </tbody>
+      </table>
+    );
+  }
+}
+
+export default MembersTable;
