+++ client/app/bundles/Events/components/Spinner.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const Spinner = (props) => (
-  <i className="fa fa-spinner fa-spin" />
+const Spinner = ({ classes }) => (
+  <i className={`fa fa-spinner fa-spin ${classes}`} />
 );
 
 export default Spinner;
